+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Added `mypy` and `pylint` at build time
 
+### Changed
+- Try to import `Retry` directly from `urllib3`, falling back to `requests.packages.urllib3` if needed.
+
 ## [0.5.5] - 2026-04-14
 
 ### Changed